Feature/survey

+++ src/main/java/com/example/spring/dto/QuestionOptionDTO.java
@@ -2,9 +2,10 @@ package com.example.spring.dto;
 
 import lombok.Data;
 
+
 @Data
 public class QuestionOptionDTO {
-    private int questionOptionId;
+    private Long questionOptionId;
     private String content;
     private double score;
